+++ src/components/ui/FlatButton.tsx
@@ -2,7 +2,7 @@
 import { ReactNode } from 'react';
 import { Pressable, Text, View } from 'react-native';
 
-import { buttonStyles } from '../../constants/styles';
+import { buttonStyles } from "../../constants/styles/buttonStyles";
 
 function FlatButton(props: { children: ReactNode, onPress: () => void }) {
   return (
