+++ src/components/ui/LiftHistory/LiftChart.tsx
@@ -1,7 +1,7 @@
 import { Dimensions, Text, View } from "react-native";
 import { LineChart } from "react-native-chart-kit";
-import { ScreenStyle } from "../../../constants/styles";
-import { Lift } from "../../userStats/UserStatsTypes";
+import { ScreenStyle } from "../../../constants/styles/screenStyles";
+import { Lift } from "../../../context/userStats/UserStatsTypes";
 import { useEffect, useState } from "react";
 import LoadingOverlay from "../../../screens/LoadingOverlay";
 import { Colors } from "../../../constants/colors";
@@ -28,8 +28,6 @@ export const LiftChart = ({ lifts, chartType }: GraphProps) => {
 	const liftData = lifts.map((lift) => lift[liftWeightKey] as number);
 
 	const labels = lifts.map((lift) => {
-
-		console.log(`timestampsdmf = ${lift.timestamp.toDate()}`)
 		const date = lift.timestamp.toDate(); // Converts Firestore Timestamp to JavaScript Date
 		const dateString = `${date.getDate()}/${date.getMonth()}`
 		return dateString
