+++ src/components/userStats/LiftsForm.tsx
@@ -5,7 +5,8 @@ import { Colors } from "../../constants/colors";
 import Button from "../ui/Button";
 import { AuthContext } from "../../context/auth/AuthProvider";
 import { UserStatsContext } from "../../context/userStats/UserStatsProvider";
-import { ScreenStyle } from "../../constants/styles";
+import { ScreenStyle } from "../../constants/styles/screenStyles";
+import { contentStyle } from "../../constants/styles/contentStyles";
 
 export const LiftsForm = () => {
 	const { user } = useContext(AuthContext);
@@ -54,7 +55,7 @@ export const LiftsForm = () => {
 	}
 
 	return (
-		<View style={ScreenStyle.updateLiftsContent}>
+		<View style={contentStyle.updateLiftsContent}>
 			<Input
 				label="Weight (kg)"
 				keyboardType="number-pad"
