+++ src/screens/friends/FriendsFeedView.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import { View } from 'react-native';
+
+export default function FriendsFeedView () {
+    return (
+        <View>
+        </View>
+    );
+}
