+++ src/screens/liftStats/LiftHistoryScreen.tsx
@@ -5,7 +5,7 @@ import { AuthContext } from "../../context/auth/AuthProvider";
 import { UserStatsContext } from "../../context/userStats/UserStatsProvider";
 
 import { liftHistoryList } from "../../components/ui/LiftHistory/LiftHistoryList";
-import { ScreenStyle } from "../../constants/styles";
+import { ScreenStyle } from "../../constants/styles/screenStyles";
 
 function LiftHistoryScreen() {
 	const { fetchAllLifts, allLifts, liftsAdded } = useContext(UserStatsContext);
